<fix> Fix duplicated request in post page SSR
